it works
i have no idea if it is correct but it works. You can create a simulation, set particle positions params etc and then simulate a step.

+++ src/types/simulationTypes.h
@@ -5,7 +5,7 @@
 // This file should contain only simulation-related data structures. GUI structs go somewhere else.
 
 struct GridData {
-    unsigned int* particle_hashes;
+    unsigned long long* particle_hashes;
     unsigned int* particle_indices;
     unsigned int* cell_starts;
     unsigned int* cell_ends;
@@ -14,7 +14,7 @@ struct GridData {
     float invGridCellSize;
     float4 domainMin;
     int4 gridDimensions;
-    unsigned int numGridCells; // WARNING: Can overflow for large 4D grids
+    unsigned long long numGridCells; // WARNING: Can overflow for large 4D grids
 };
 
 struct SimulationParams {
